Moved GraphQLAuthorizeAttribute and AuthorizationMetadataExtensions to new Authorization package
