+++ GraphQL.Server.sln
@@ -77,6 +77,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ApiApprovalTests", "tests\A
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "All", "src\All\All.csproj", "{B39BC064-B96A-442D-89A0-79B750F88ADD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Authorization", "src\Authorization\Authorization.csproj", "{152E46DC-7225-4676-8942-14E2D51819B6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -163,6 +165,10 @@ Global
 		{B39BC064-B96A-442D-89A0-79B750F88ADD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B39BC064-B96A-442D-89A0-79B750F88ADD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B39BC064-B96A-442D-89A0-79B750F88ADD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{152E46DC-7225-4676-8942-14E2D51819B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{152E46DC-7225-4676-8942-14E2D51819B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{152E46DC-7225-4676-8942-14E2D51819B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{152E46DC-7225-4676-8942-14E2D51819B6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
