+++ src/Authorization/Authorization.csproj
@@ -2,13 +2,12 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <Description>Integration of GraphQL.NET validation subsystem into ASP.NET Core</Description>
+    <Description>Integration of GraphQL.NET validation subsystem</Description>
     <PackageTags>GraphQL;authentication;authorization;validation</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-    <ProjectReference Include="..\Authorization\Authorization.csproj" />
+    <ProjectReference Include="..\Core\Core.csproj" />
   </ItemGroup>
 
 </Project>
